requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.6.15
et_xmlfile==2.0.0
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
openpyxl==3.1.5
propcache==0.3.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.1.1
typing_extensions==4.14.0
yarl==1.20.1
